requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-prometheus==2.2.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
django-filter==22.1
django-redis==5.2.0
django-simple-history==3.1.1
django-taggit==3.0.0
django-storages==1.13.1
rest-framework-roles==0.4.1
flake8==4.0.1
gunicorn==20.1.0
jsonschema==4.16.0
mccabe==0.6.1
prometheus-client==0.14.1
psutil==5.9.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
psutil==5.9.1
pyyaml==6.0.0
uritemplate==4.1.1
celery==5.2.7
django-celery-results==2.4.0
redis==4.3.4
slack_sdk==3.18.1
boto3==1.24.61
moto==4.0.1
moviepy==1.0.3